pyproject.toml
[project]
name = "downloader"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "VitorAndreSilva", email = "alunovitorandre@gmail.com"},
]
dependencies = ["django>=5.2.7", "djangorestframework>=3.16.1", "playwright>=1.56.0", "requests>=2.32.5", "yt-dlp>=2025.11.12"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
